requirements.txt
# pip freeze > requirements.txt

astroid==2.0.1
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
DateTime==4.2
Django==2.0.7
idna==2.7
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==2.2.2
mccabe==0.6.1
nltk==3.3
numpy==1.14.5
oauthlib==2.1.0
pandas==0.23.3
pylint==2.0.0
pyparsing==2.2.0
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
sklearn==0.0
textblob==0.15.1
tweepy==3.6.0
typed-ast==1.1.0
urllib3==1.23
wrapt==1.10.11
zope.interface==4.5.0
